# [clflush] Enable x86 cpu cache flush