--- a/python/tvm/runtime/module.py
+++ b/python/tvm/runtime/module.py
@@ -163,7 +163,7 @@ class Module(object):
         """
         _ffi_api.ModuleSaveToFile(self, file_name, fmt)
 
-    def time_evaluator(self, func_name, ctx, number=10, repeat=1, min_repeat_ms=0):
+    def time_evaluator(self, func_name, ctx, number=10, repeat=1, min_repeat_ms=0, f_prepare=None):
         """Get an evaluator that measures time cost of running function.
 
         Parameters
@@ -207,7 +207,7 @@ class Module(object):
         try:
             feval = _ffi_api.RPCTimeEvaluator(
                 self, func_name, ctx.device_type, ctx.device_id,
-                number, repeat, min_repeat_ms)
+                number, repeat, min_repeat_ms, f_prepare)
 
             def evaluator(*args):
                 """Internal wrapped evaluator."""