… tests/python/unittest/test_autotvm_measure.py
… tutorials/autotvm/tune_relay_x86.py