@tqchen 2020-06-24
cc @jwfromm @junrushao1994 @merrymercy @yidawang @jcf94 for quick discussion