@yidawang 2020-06-24
I agree that the cache flush mechanism is useful in getting preciser measurement. It would be great if @FrozenGene can provide some experimental data to further assure.

I vote for folding cache flushing factor into time_evaluator for succinctness. And making it more configurable and generic sounds good to me.