@FrozenGene 2020-06-28
I was at vocation several days ago, sorry for replying a little late. For this, I also want to do a quick discussion. For single op / or Ansor's single subgraph benchmark, we don't want to turn on cache flush to unleash maximum performance, however for network benchmark, we want to do it because of the reason mentioned. So maybe A1  is easier to complete this task if everyone agree we turn off cache flush when we do single op benchmark?

I agree we should provide one api for this as if we want to support remote devices (like arm), environment values can not control it on our local machine.