@tqchen 2020-06-28
I agree that A1 is more modular, @yidawang do you have strong preference on A0?