@tqchen 2020-06-29
How about we go with A1 for now, @FrozenGene can you update this PR to A1?