@FrozenGene 2020-06-30
> How about we go with A1 for now, @FrozenGene can you update this PR to A1?

OK. I will implement this PR to A1 next.