@FrozenGene 2020-06-30
> How about we go with A1 for now, @FrozenGene can you update this PR to A1?

@tqchen If we use
 ```
# cache flush packed is a packed func that performs the cpu cache flush
cache_flush_packed = remote.get_function("cpu_cache_flush")(args=begin=1)
# fprepare is a callback that will be called before the evaluation, it takes in args as arguments. 
f = mod.time_evaluator("myfunc", repeat=10, fprepare=cache_flush_packed)
```
We will meet `Cannot pass type FunctionHandle as an argument to the remote` error. Do you have any good suggestion about it?